+++ CHANGELOG.md
@@ -15,3 +15,9 @@
 - **What I built:** Successfully fetching from Indiehackers using Algolia search. Getting primary information like bio, name, and email.
 - **Why it matters:** Cracking IndieHackers gives us a nice starting point to expand the thoroughness of our leads through either other platforms or deeper IndieHackers searches.
 - **Next:** Explore what else can be scraped from IndieHackers briefly. If nothing, move to correlating information to other platforms like Github.
+
+### 2025-05-05 | 6:00 PM 🧭 Daily Log
+
+- **What I built:** A BFS building of subsections for username prefixes to get over Algolia 1000 search limit and proper polite throttling
+- **Why it matters:** Getting usernames from IndieHackers means we have a big stem of data to start from and can move on to other platforms. This is a big starting step.
+- **Next:** Explore all users with prefixes and get lead information for agentic system and other relations.